langchain-agent: fix issue when using OnnxBertBiEncoder models in rag scenarios #8370
