+++ extensions-support/langchain4j/deployment/src/main/java/org/apache/camel/quarkus/component/support/langchain4j/deployment/SupportLangchain4jProcessor.java
@@ -16,12 +16,16 @@
  */
 package org.apache.camel.quarkus.component.support.langchain4j.deployment;
 
+import java.io.IOException;
+import java.net.URL;
 import java.util.Collection;
+import java.util.Enumeration;
 import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import ai.djl.util.NativeResource;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
@@ -29,17 +33,21 @@ import dev.langchain4j.agent.tool.Tool;
 import dev.langchain4j.guardrail.InputGuardrail;
 import dev.langchain4j.guardrail.JsonExtractorOutputGuardrail;
 import dev.langchain4j.guardrail.OutputGuardrail;
+import dev.langchain4j.model.embedding.onnx.AbstractInProcessEmbeddingModel;
 import dev.langchain4j.service.MemoryId;
 import dev.langchain4j.service.SystemMessage;
 import dev.langchain4j.service.UserMessage;
 import dev.langchain4j.service.V;
+import io.quarkus.bootstrap.classloading.QuarkusClassLoader;
 import io.quarkus.bootstrap.model.ApplicationModel;
 import io.quarkus.deployment.annotations.BuildProducer;
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.annotations.BuildSteps;
 import io.quarkus.deployment.builditem.CombinedIndexBuildItem;
 import io.quarkus.deployment.builditem.IndexDependencyBuildItem;
+import io.quarkus.deployment.builditem.nativeimage.JniRuntimeAccessBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.NativeImageProxyDefinitionBuildItem;
+import io.quarkus.deployment.builditem.nativeimage.NativeImageResourceBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.NativeImageResourcePatternsBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.ReflectiveClassBuildItem;
 import io.quarkus.deployment.builditem.nativeimage.RuntimeInitializedClassBuildItem;
@@ -47,7 +55,9 @@ import io.quarkus.deployment.builditem.nativeimage.ServiceProviderBuildItem;
 import io.quarkus.deployment.pkg.builditem.CurateOutcomeBuildItem;
 import io.quarkus.deployment.pkg.steps.NativeOrNativeSourcesBuild;
 import io.quarkus.maven.dependency.ResolvedDependency;
+import io.smallrye.common.os.OS;
 import opennlp.tools.sentdetect.SentenceDetectorFactory;
+import org.apache.camel.util.ObjectHelper;
 import org.jboss.jandex.AnnotationInstance;
 import org.jboss.jandex.AnnotationTarget;
 import org.jboss.jandex.ClassInfo;
@@ -230,5 +240,168 @@ class SupportLangchain4jProcessor {
     static Stream<ClassInfo> langChain4jTypesStream(Collection<ClassInfo> classes) {
         return classes.stream()
                 .filter(classInfo -> classInfo.name().toString().startsWith("dev.langchain4j"));
+    }
+
+    @BuildStep(onlyIf = NativeOrNativeSourcesBuild.class)
+    void indexDjlDependencies(
+            CurateOutcomeBuildItem curateOutcome,
+            BuildProducer<IndexDependencyBuildItem> indexDependency) {
+
+        ApplicationModel applicationModel = curateOutcome.getApplicationModel();
+        for (ResolvedDependency dependency : applicationModel.getDependencies()) {
+            String groupId = dependency.getGroupId();
+            String artifactId = dependency.getArtifactId();
+
+            // Index ai.djl dependencies
+            if (groupId.startsWith("ai.djl")) {
+                indexDependency.produce(new IndexDependencyBuildItem(groupId, artifactId));
+            }
+        }
+    }
+
+    @BuildStep(onlyIf = NativeOrNativeSourcesBuild.class)
+    void nativeImageResources(
+            CurateOutcomeBuildItem curateOutcome,
+            BuildProducer<NativeImageResourcePatternsBuildItem> nativeImageResourcePattern,
+            BuildProducer<NativeImageResourceBuildItem> nativeImageResource) {
+
+        // The supported platforms for native binaries within ai.djl.huggingface:tokenizers & com.microsoft.onnxruntime:onnxruntime
+        String os = switch (OS.current()) {
+        case LINUX -> "linux";
+        case WINDOWS -> "win";
+        case MAC -> "osx";
+        default -> null;
+        };
+
+        // Search for native binaries for the target platform and add them to the native image
+        String osArch = System.getProperty("os.arch");
+        if (ObjectHelper.isNotEmpty(os)) {
+            Stream.of(AI_LIBRARY.values()).forEach(aiLibrary -> {
+                try {
+                    String binaryPath = aiLibrary.getBinaryPath(os);
+                    Enumeration<URL> resources = Thread.currentThread()
+                            .getContextClassLoader()
+                            .getResources(binaryPath);
+                    if (resources.hasMoreElements()) {
+                        nativeImageResourcePattern
+                                .produce(NativeImageResourcePatternsBuildItem.builder()
+                                        .includeGlobs(binaryPath + "/**").build());
+                    } else {
+                        throw new RuntimeException(
+                                "Failed to find required native binaries on classpath at %s".formatted(binaryPath));
+                    }
+                } catch (IOException e) {
+                    throw new RuntimeException(e);
+                }
+            });
+        } else {
+            throw new RuntimeException("%s does not support platform %s-%s".formatted("SupportLangchain4j", os, osArch));
+        }
+
+        nativeImageResource.produce(new NativeImageResourceBuildItem("native/lib/tokenizers.properties"));
+
+        for (ResolvedDependency dependency : curateOutcome.getApplicationModel().getDependencies()) {
+            String artifactId = dependency.getArtifactId();
+
+            // Native image resources for embeddings tokenizer metadata
+            String prefix = "langchain4j-embeddings-";
+            if (artifactId.startsWith(prefix)) {
+                String embeddingLibType = artifactId.substring(prefix.length());
+                nativeImageResource.produce(new NativeImageResourceBuildItem(embeddingLibType + ".onnx"));
+                nativeImageResource.produce(new NativeImageResourceBuildItem(embeddingLibType + "-tokenizer.json"));
+            }
+        }
+    }
+
+    @BuildStep(onlyIf = NativeOrNativeSourcesBuild.class)
+    void jniRuntimeSupport(
+            BuildProducer<ReflectiveClassBuildItem> reflectiveClass,
+            BuildProducer<JniRuntimeAccessBuildItem> jniRuntimeAccess) {
+
+        String[] jniReflectiveClasses = new String[] {
+                "ai.onnxruntime.TensorInfo",
+                "ai.onnxruntime.SequenceInfo",
+                "ai.onnxruntime.MapInfo",
+                "ai.onnxruntime.OrtException",
+                "ai.onnxruntime.OnnxSparseTensor",
+                "ai.djl.huggingface.tokenizers.jni.CharSpan",
+                "[Lai.djl.huggingface.tokenizers.jni.CharSpan;",
+                "[Lai.onnxruntime.OnnxValue;",
+                "[Ljava.lang.String;",
+                "ai.onnxruntime.OnnxTensor",
+                "ai.onnxruntime.OnnxValue",
+                "ai.onnxruntime.TensorInfo",
+                "ai.onnxruntime.OnnxTensor",
+                "java.lang.Boolean",
+                "java.lang.String"
+        };
+
+        reflectiveClass.produce(
+                ReflectiveClassBuildItem.builder(jniReflectiveClasses)
+                        .fields()
+                        .methods()
+                        .build());
+
+        jniRuntimeAccess.produce(
+                new JniRuntimeAccessBuildItem(true, true, true, jniReflectiveClasses));
+
+        reflectiveClass.produce(ReflectiveClassBuildItem.builder("opennlp.tools.sentdetect.SentenceDetectorFactory").build());
+    }
+
+    @BuildStep(onlyIf = NativeOrNativeSourcesBuild.class)
+    void runtimeInitializedClasses(
+            CombinedIndexBuildItem combinedIndex,
+            BuildProducer<RuntimeInitializedClassBuildItem> runtimeInitializedClass) {
+
+        IndexView index = combinedIndex.getIndex();
+
+        // Due to JNI usage the following types must be initialized at runtime
+        index.getAllKnownSubclasses(AbstractInProcessEmbeddingModel.class)
+                .stream()
+                .map(classInfo -> classInfo.name().toString())
+                .map(RuntimeInitializedClassBuildItem::new)
+                .forEach(runtimeInitializedClass::produce);
+
+        index.getAllKnownSubclasses(NativeResource.class)
+                .stream()
+                .map(classInfo -> classInfo.name().toString())
+                .map(RuntimeInitializedClassBuildItem::new)
+                .forEach(runtimeInitializedClass::produce);
+
+        Stream.of("ai.djl.huggingface.tokenizers.jni.TokenizersLibrary",
+                "ai.djl.huggingface.tokenizers.jni.LibUtils",
+                "ai.djl.util.Platform",
+                "ai.onnxruntime.OrtEnvironment",
+                "ai.onnxruntime.OnnxRuntime",
+                "ai.onnxruntime.OnnxTensorLike",
+                "ai.onnxruntime.OrtSession$SessionOptions")
+                .filter(QuarkusClassLoader::isClassPresentAtRuntime)
+                .map(RuntimeInitializedClassBuildItem::new)
+                .forEach(runtimeInitializedClass::produce);
+    }
+
+    enum AI_LIBRARY {
+        AI_DJL_HUGGINGFACE_TOKENIZERS("ai/onnxruntime/native/%s-%s", "x64"),
+        COM_MICROSOFT_ONNXRUNTIME("native/lib/%s-%s/cpu", "x86_64");
+
+        private final String binaryPathTemplate;
+        private final String x86Arch;
+
+        AI_LIBRARY(String binaryPathTemplate, String x86Arch) {
+            this.binaryPathTemplate = binaryPathTemplate;
+            this.x86Arch = x86Arch;
+        }
+
+        public String getBinaryPath(String os) {
+            return binaryPathTemplate.formatted(os, getBinaryArch());
+        }
+
+        public String getBinaryArch() {
+            String arch = System.getProperty("os.arch");
+            if (arch.equals("amd64")) {
+                return this.x86Arch;
+            }
+            return arch;
+        }
     }
 }
