+++ extensions-support/langchain4j/runtime/pom.xml
@@ -46,6 +46,10 @@
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-core</artifactId>
         </dependency>
+        <dependency>
+            <groupId>dev.langchain4j</groupId>
+            <artifactId>langchain4j-embeddings</artifactId>
+        </dependency>
     </dependencies>
 
     <build>
