+++ extensions/langchain4j-embeddings/runtime/pom.xml
@@ -49,10 +49,6 @@
             <groupId>org.apache.camel</groupId>
             <artifactId>camel-langchain4j-embeddings</artifactId>
         </dependency>
-        <dependency>
-            <groupId>dev.langchain4j</groupId>
-            <artifactId>langchain4j-embeddings</artifactId>
-        </dependency>
     </dependencies>
 
     <build>
