+++ integration-tests/langchain4j-agent/pom.xml
@@ -59,6 +59,10 @@
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-ollama</artifactId>
         </dependency>
+        <dependency>
+            <groupId>dev.langchain4j</groupId>
+            <artifactId>langchain4j-embeddings-all-minilm-l6-v2</artifactId>
+        </dependency>
 
         <!-- test dependencies -->
         <dependency>
