+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-3dbd2721-ecf7-4960-a6d9-532629c974d1.json
@@ -1,5 +1,5 @@
 {
-  "id" : "b9e443eb-f523-43c5-b15a-5b481b00ad7e",
+  "id" : "3dbd2721-ecf7-4960-a6d9-532629c974d1",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-01-21T08:00:56.666262664Z\",\"message\":{\"role\":\"assistant\",\"content\":\"The name of user ID 123 is John Doe.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":839226209,\"load_duration\":45703671,\"prompt_eval_count\":230,\"prompt_eval_duration\":343699976,\"eval_count\":12,\"eval_duration\":440847701}",
+    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-03-03T15:25:15.36019929Z\",\"message\":{\"role\":\"assistant\",\"content\":\"John Doe\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":873488110,\"load_duration\":53320906,\"prompt_eval_count\":248,\"prompt_eval_duration\":704054631,\"eval_count\":3,\"eval_duration\":110231366}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:00:56 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:25:15 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "b9e443eb-f523-43c5-b15a-5b481b00ad7e",
+  "uuid" : "3dbd2721-ecf7-4960-a6d9-532629c974d1",
   "persistent" : true,
   "insertionIndex" : 19
 }
