+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-46dd74e1-aa17-497e-bad5-fc74ee8d7959.json
@@ -1,0 +1,24 @@
+{
+  "id" : "46dd74e1-aa17-497e-bad5-fc74ee8d7959",
+  "name" : "api_chat",
+  "request" : {
+    "url" : "/api/chat",
+    "method" : "POST",
+    "bodyPatterns" : [ {
+      "equalToJson" : "{\n  \"model\" : \"orca-mini\",\n  \"messages\" : [ {\n    \"role\" : \"user\",\n    \"content\" : \"Describe the Miles of Camels Car Rental cancellations policy for cancelling 24 hours before pickup. What is the refund amount?\\n\\nAnswer using the following information:\\nMiles of Camels Car Rental - Company Information\\n\\nBUSINESS HOURS:\\nMonday-Friday: 8:00 AM - 6:00 PM\\nSaturday: 9:00 AM - 4:00 PM\\nSunday: Closed\\n\\nRENTAL AGREEMENT\\n- This agreement is between Miles of Camels Car Rental (\\\"Company\\\") and the customer (\\\"Renter\\\").\\n\\nCANCELLATION POLICY\\n- Cancellations made 24 hours before pickup: Full refund\\n- Cancellations made 12-24 hours before pickup: 50% refund\\n- Cancellations made less than 12 hours before pickup: No refund\\n\\nVEHICLE RETURN\\n- Vehicles must be returned with the same fuel level as at pickup.\\n- Late returns incur a fee of $25 per hour or fraction thereof.\\n\\nDAMAGE POLICY\\n- Minor damages under $200: Covered by insurance\\n- Major damages over $200: Customer responsibility\"\n  } ],\n  \"options\" : {\n    \"temperature\" : 0.3,\n    \"stop\" : [ ]\n  },\n  \"stream\" : false,\n  \"tools\" : [ ]\n}",
+      "ignoreArrayOrder" : true,
+      "ignoreExtraElements" : false
+    } ]
+  },
+  "response" : {
+    "status" : 200,
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:25:09.879393101Z\",\"message\":{\"role\":\"assistant\",\"content\":\" According to Miles of Camels Car Rental's cancellations policy, if you cancel your rental 24 hours before pickup, you will receive a full refund. If you cancel between 12 and 24 hours before pickup, you will receive a 50% refund. If you cancel less than 12 hours before pickup, there will be no refund.\\n\\nAdditionally, according to the company's vehicle return policy, vehicles must be returned with the same fuel level as at pickup. Late returns incur a fee of $25 per hour or fraction thereof.\\n\\nFinally, according to Miles of Camels Car Rental's damage policy, major damages exceeding $200 are the responsibility of the customer, while minor damages under $200 are covered by insurance.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":16373127648,\"load_duration\":12717353,\"prompt_eval_count\":314,\"prompt_eval_duration\":6552571248,\"eval_count\":166,\"eval_duration\":9763683602}",
+    "headers" : {
+      "Date" : "Tue, 03 Mar 2026 15:25:09 GMT",
+      "Content-Type" : "application/json; charset=utf-8"
+    }
+  },
+  "uuid" : "46dd74e1-aa17-497e-bad5-fc74ee8d7959",
+  "persistent" : true,
+  "insertionIndex" : 21
+}
