+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-50282a37-becc-455a-829b-8c95dbfeac83.json
@@ -1,5 +1,5 @@
 {
-  "id" : "c285552d-f27e-4cf3-a942-ff92a666e118",
+  "id" : "50282a37-becc-455a-829b-8c95dbfeac83",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-01-21T08:00:55.817907064Z\",\"message\":{\"role\":\"assistant\",\"content\":\"\",\"tool_calls\":[{\"id\":\"call_pnsg7jwe\",\"function\":{\"index\":0,\"name\":\"QueryUserDatabaseByUserID\",\"arguments\":{\"userId\":123}}}]},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":4644671100,\"load_duration\":1450597042,\"prompt_eval_count\":182,\"prompt_eval_duration\":2310269399,\"eval_count\":23,\"eval_duration\":872125825}",
+    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-03-03T15:25:14.47274051Z\",\"message\":{\"role\":\"assistant\",\"content\":\"\",\"tool_calls\":[{\"id\":\"call_w3upfd1y\",\"function\":{\"index\":0,\"name\":\"QueryUserDatabaseByUserID\",\"arguments\":{\"userId\":123}}}]},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":4472695102,\"load_duration\":43611230,\"prompt_eval_count\":200,\"prompt_eval_duration\":3210769739,\"eval_count\":23,\"eval_duration\":1200591755}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:00:55 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:25:14 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "c285552d-f27e-4cf3-a942-ff92a666e118",
+  "uuid" : "50282a37-becc-455a-829b-8c95dbfeac83",
   "persistent" : true,
   "insertionIndex" : 20
 }
