+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-5b13dcfa-8f76-45b3-a149-6c43182869dd.json
@@ -1,5 +1,5 @@
 {
-  "id" : "2b0b28a2-8580-47f8-a2a9-d5f1a315c7ba",
+  "id" : "5b13dcfa-8f76-45b3-a149-6c43182869dd",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-01-21T08:01:20.936033958Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Sure, here's an example JSON object for John Doe:\\n\\n```json\\n{\\n \\\"name\\\": \\\"John Doe\\\",\\n \\\"description\\\": \\\"An example person with a unique name and a brief description\\\"\\n}\\n```\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":2504619772,\"load_duration\":11930628,\"prompt_eval_count\":60,\"prompt_eval_duration\":464597196,\"eval_count\":53,\"eval_duration\":2018296423}",
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:25:47.817132698Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Sure, here's an example JSON object for John Doe:\\n\\n```json\\n{\\n \\\"name\\\": \\\"John Doe\\\",\\n \\\"description\\\": \\\"An example person with a unique name and a brief description\\\"\\n}\\n```\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":3933350526,\"load_duration\":11705957,\"prompt_eval_count\":60,\"prompt_eval_duration\":1173751415,\"eval_count\":53,\"eval_duration\":2733823806}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:20 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:25:47 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "2b0b28a2-8580-47f8-a2a9-d5f1a315c7ba",
+  "uuid" : "5b13dcfa-8f76-45b3-a149-6c43182869dd",
   "persistent" : true,
   "insertionIndex" : 15
 }
