+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-63b71ca8-026b-471b-a34a-da7660be2631.json
@@ -1,5 +1,5 @@
 {
-  "id" : "946ce08a-8829-4ffa-b425-6609cd8413e6",
+  "id" : "63b71ca8-026b-471b-a34a-da7660be2631",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-01-21T08:01:47.110185749Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Hello, Alice. How can I assist you today?\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":613502063,\"load_duration\":11960283,\"prompt_eval_count\":46,\"prompt_eval_duration\":149859087,\"eval_count\":12,\"eval_duration\":448584137}",
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:26:27.290839245Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Hello, Alice. How can I assist you today?\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":932736372,\"load_duration\":11343198,\"prompt_eval_count\":46,\"prompt_eval_duration\":341050678,\"eval_count\":12,\"eval_duration\":576140320}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:47 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:26:27 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "946ce08a-8829-4ffa-b425-6609cd8413e6",
+  "uuid" : "63b71ca8-026b-471b-a34a-da7660be2631",
   "persistent" : true,
   "scenarioName" : "scenario-2-api-chat",
   "requiredScenarioState" : "scenario-2-api-chat-3",
