+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-65a5a977-8a4c-4c28-aff7-f4dc7500df43.json
@@ -1,5 +1,5 @@
 {
-  "id" : "334a44c3-a293-4f2e-a358-d2de41f194d0",
+  "id" : "65a5a977-8a4c-4c28-aff7-f4dc7500df43",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,16 +12,16 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-01-21T08:00:57.311525014Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Hello, Alice. How can I assist you today?\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":622877906,\"load_duration\":11845848,\"prompt_eval_count\":46,\"prompt_eval_duration\":188038439,\"eval_count\":12,\"eval_duration\":419968471}",
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:26:13.220162181Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Hello, Alice. How can I assist you today?\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":900089323,\"load_duration\":9450899,\"prompt_eval_count\":46,\"prompt_eval_duration\":319079740,\"eval_count\":12,\"eval_duration\":567495610}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:00:57 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:26:13 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "334a44c3-a293-4f2e-a358-d2de41f194d0",
+  "uuid" : "65a5a977-8a4c-4c28-aff7-f4dc7500df43",
   "persistent" : true,
   "scenarioName" : "scenario-2-api-chat",
-  "requiredScenarioState" : "Started",
-  "newScenarioState" : "scenario-2-api-chat-2",
-  "insertionIndex" : 18
+  "requiredScenarioState" : "scenario-2-api-chat-2",
+  "newScenarioState" : "scenario-2-api-chat-3",
+  "insertionIndex" : 8
 }
