+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-68c8507d-160c-4641-9a22-84bbe4806890.json
@@ -1,0 +1,24 @@
+{
+  "id" : "68c8507d-160c-4641-9a22-84bbe4806890",
+  "name" : "api_chat",
+  "request" : {
+    "url" : "/api/chat",
+    "method" : "POST",
+    "bodyPatterns" : [ {
+      "equalToJson" : "{\n  \"model\" : \"orca-mini\",\n  \"messages\" : [ {\n    \"role\" : \"user\",\n    \"content\" : \"What is Apache Camel?\"\n  } ],\n  \"options\" : {\n    \"temperature\" : 0.3,\n    \"stop\" : [ ]\n  },\n  \"stream\" : false,\n  \"tools\" : [ ]\n}",
+      "ignoreArrayOrder" : true,
+      "ignoreExtraElements" : false
+    } ]
+  },
+  "response" : {
+    "status" : 200,
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:25:21.603626775Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Apache Camel is an open-source framework for building and managing business processes that run on top of Camel Transport. It provides a modular, pluggable architecture that allows you to easily integrate Camel with other technologies such as Java, Spring, Hibernate, and more. Camel can be used to automate tasks such as message routing, load balancing, and message processing. It is commonly used in building distributed systems, IoT applications, and web services.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":5279325682,\"load_duration\":13812216,\"prompt_eval_count\":46,\"prompt_eval_duration\":327187702,\"eval_count\":95,\"eval_duration\":4915431904}",
+    "headers" : {
+      "Date" : "Tue, 03 Mar 2026 15:25:21 GMT",
+      "Content-Type" : "application/json; charset=utf-8"
+    }
+  },
+  "uuid" : "68c8507d-160c-4641-9a22-84bbe4806890",
+  "persistent" : true,
+  "insertionIndex" : 17
+}
