+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-6b1b3f2e-5aa1-4c9b-88d3-f24a10590c6f.json
@@ -1,5 +1,5 @@
 {
-  "id" : "d4f1693c-88e0-456f-b650-77a6bf6e543f",
+  "id" : "6b1b3f2e-5aa1-4c9b-88d3-f24a10590c6f",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,16 +12,16 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-01-21T08:01:37.577431121Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Hello, Alice. How can I assist you today?\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":596937019,\"load_duration\":11922904,\"prompt_eval_count\":46,\"prompt_eval_duration\":152871930,\"eval_count\":12,\"eval_duration\":428769988}",
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:25:16.303883793Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Hello, Alice. How can I assist you today?\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":920820435,\"load_duration\":10510752,\"prompt_eval_count\":46,\"prompt_eval_duration\":343358491,\"eval_count\":12,\"eval_duration\":563204651}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:37 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:25:16 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "d4f1693c-88e0-456f-b650-77a6bf6e543f",
+  "uuid" : "6b1b3f2e-5aa1-4c9b-88d3-f24a10590c6f",
   "persistent" : true,
   "scenarioName" : "scenario-2-api-chat",
-  "requiredScenarioState" : "scenario-2-api-chat-2",
-  "newScenarioState" : "scenario-2-api-chat-3",
-  "insertionIndex" : 8
+  "requiredScenarioState" : "Started",
+  "newScenarioState" : "scenario-2-api-chat-2",
+  "insertionIndex" : 18
 }
