--- integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-6fdb3b04-ba4c-4add-956c-e1c8ec82f799.json
@@ -1,24 +1,0 @@
-{
-  "id" : "6fdb3b04-ba4c-4add-956c-e1c8ec82f799",
-  "name" : "api_chat",
-  "request" : {
-    "url" : "/api/chat",
-    "method" : "POST",
-    "bodyPatterns" : [ {
-      "equalToJson" : "{\n  \"model\" : \"orca-mini\",\n  \"messages\" : [ {\n    \"role\" : \"user\",\n    \"content\" : \"Return an example JSON object about a person named 'John Doe' with the fields age and description\"\n  } ],\n  \"options\" : {\n    \"temperature\" : 0.3,\n    \"stop\" : [ ]\n  },\n  \"stream\" : false,\n  \"tools\" : [ ]\n}",
-      "ignoreArrayOrder" : true,
-      "ignoreExtraElements" : false
-    } ]
-  },
-  "response" : {
-    "status" : 200,
-    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-01-21T08:01:41.39971525Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Sure, here's an example JSON object for John Doe with his age and description:\\n\\n```json\\n{\\n \\\"name\\\": \\\"John Doe\\\",\\n \\\"age\\\": 30,\\n \\\"description\\\": \\\"John Doe is a software engineer with over five years of experience. He enjoys hiking and playing guitar in his free time.\\\"\\n}\\n```\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":3805629638,\"load_duration\":13049337,\"prompt_eval_count\":60,\"prompt_eval_duration\":420573116,\"eval_count\":79,\"eval_duration\":3356822651}",
-    "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:41 GMT",
-      "Content-Type" : "application/json; charset=utf-8"
-    }
-  },
-  "uuid" : "6fdb3b04-ba4c-4add-956c-e1c8ec82f799",
-  "persistent" : true,
-  "insertionIndex" : 7
-}
