+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-84aa25c6-6069-4780-b9d4-4068a07c101a.json
@@ -1,0 +1,24 @@
+{
+  "id" : "84aa25c6-6069-4780-b9d4-4068a07c101a",
+  "name" : "api_chat",
+  "request" : {
+    "url" : "/api/chat",
+    "method" : "POST",
+    "bodyPatterns" : [ {
+      "equalToJson" : "{\n  \"model\" : \"orca-mini\",\n  \"messages\" : [ {\n    \"role\" : \"user\",\n    \"content\" : \"Return an example JSON object about a person named 'John Doe' with the fields age and description\"\n  } ],\n  \"options\" : {\n    \"temperature\" : 0.3,\n    \"stop\" : [ ]\n  },\n  \"stream\" : false,\n  \"tools\" : [ ]\n}",
+      "ignoreArrayOrder" : true,
+      "ignoreExtraElements" : false
+    } ]
+  },
+  "response" : {
+    "status" : 200,
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:26:18.135421673Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Sure, here's an example JSON object for John Doe with the fields age and description:\\n\\n```json\\n{\\n \\\"name\\\": \\\"John Doe\\\",\\n \\\"age\\\": 30,\\n \\\"description\\\": \\\"John Doe is a professional software developer with over 5 years of experience. He enjoys hiking and playing guitar in his free time.\\\"\\n}\\n```\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":4897979696,\"load_duration\":12229479,\"prompt_eval_count\":60,\"prompt_eval_duration\":649672363,\"eval_count\":82,\"eval_duration\":4214259782}",
+    "headers" : {
+      "Date" : "Tue, 03 Mar 2026 15:26:18 GMT",
+      "Content-Type" : "application/json; charset=utf-8"
+    }
+  },
+  "uuid" : "84aa25c6-6069-4780-b9d4-4068a07c101a",
+  "persistent" : true,
+  "insertionIndex" : 7
+}
