+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-98a9059a-eb53-4580-af3c-d32c1a30b95d.json
@@ -1,5 +1,5 @@
 {
-  "id" : "3cd813bc-486c-48ca-862d-5026f98fc91c",
+  "id" : "98a9059a-eb53-4580-af3c-d32c1a30b95d",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-01-21T08:01:33.067585999Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Sure, here's an example JSON object for John Doe:\\n\\n```json\\n{\\n \\\"name\\\": \\\"John Doe\\\",\\n \\\"description\\\": \\\"A person with a unique name and a brief description.\\\"\\n}\\n```\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":2473107200,\"load_duration\":12235280,\"prompt_eval_count\":90,\"prompt_eval_duration\":438293488,\"eval_count\":52,\"eval_duration\":2012512628}",
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:26:06.209091727Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Sure, here's an example JSON object for John Doe:\\n\\n```json\\n{\\n \\\"name\\\": \\\"John Doe\\\",\\n \\\"description\\\": \\\"An example person with a unique name and a brief description.\\\"\\n}\\n```\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":3900537511,\"load_duration\":9482003,\"prompt_eval_count\":90,\"prompt_eval_duration\":1148112082,\"eval_count\":53,\"eval_duration\":2728361766}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:33 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:26:06 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "3cd813bc-486c-48ca-862d-5026f98fc91c",
+  "uuid" : "98a9059a-eb53-4580-af3c-d32c1a30b95d",
   "persistent" : true,
   "insertionIndex" : 11
 }
