+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-cbafc9dd-bc75-4283-aba6-e2950556ba5f.json
@@ -1,5 +1,5 @@
 {
-  "id" : "5cf2abf3-353d-47ae-a5ef-a093c49748c8",
+  "id" : "cbafc9dd-bc75-4283-aba6-e2950556ba5f",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-01-21T08:01:35.959162737Z\",\"message\":{\"role\":\"assistant\",\"content\":\"\",\"tool_calls\":[{\"id\":\"call_ixnrpy7u\",\"function\":{\"index\":0,\"name\":\"addNumbers\",\"arguments\":{\"arg0\":10,\"arg1\":5}}}]},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":2857200262,\"load_duration\":43778600,\"prompt_eval_count\":193,\"prompt_eval_duration\":1576333093,\"eval_count\":28,\"eval_duration\":1222509432}",
+    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-03-03T15:26:10.940178285Z\",\"message\":{\"role\":\"assistant\",\"content\":\"\",\"tool_calls\":[{\"id\":\"call_nd08sj0v\",\"function\":{\"index\":0,\"name\":\"addNumbers\",\"arguments\":{\"arg0\":10,\"arg1\":5}}}]},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":4693362054,\"load_duration\":47542777,\"prompt_eval_count\":193,\"prompt_eval_duration\":3128994736,\"eval_count\":28,\"eval_duration\":1492960511}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:35 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:26:10 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "5cf2abf3-353d-47ae-a5ef-a093c49748c8",
+  "uuid" : "cbafc9dd-bc75-4283-aba6-e2950556ba5f",
   "persistent" : true,
   "insertionIndex" : 10
 }
