+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-d2554971-ae78-4b13-99c0-57def81dfb2e.json
@@ -1,5 +1,5 @@
 {
-  "id" : "45682c8e-3456-4fc9-b941-6f702f1f087a",
+  "id" : "d2554971-ae78-4b13-99c0-57def81dfb2e",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-01-21T08:01:36.945337351Z\",\"message\":{\"role\":\"assistant\",\"content\":\"The sum of 10 and 5 is **15**.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":977860704,\"load_duration\":45035949,\"prompt_eval_count\":233,\"prompt_eval_duration\":327764002,\"eval_count\":14,\"eval_duration\":595399685}",
+    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-03-03T15:26:12.287839617Z\",\"message\":{\"role\":\"assistant\",\"content\":\"The sum of 10 and 5 is **15**.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":1341044668,\"load_duration\":50634799,\"prompt_eval_count\":233,\"prompt_eval_duration\":573930922,\"eval_count\":14,\"eval_duration\":705548128}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:36 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:26:12 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "45682c8e-3456-4fc9-b941-6f702f1f087a",
+  "uuid" : "d2554971-ae78-4b13-99c0-57def81dfb2e",
   "persistent" : true,
   "insertionIndex" : 9
 }
