+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-d366f73b-f02a-4f62-ad43-836f42226e05.json
@@ -1,0 +1,24 @@
+{
+  "id" : "d366f73b-f02a-4f62-ad43-836f42226e05",
+  "name" : "api_chat",
+  "request" : {
+    "url" : "/api/chat",
+    "method" : "POST",
+    "bodyPatterns" : [ {
+      "equalToJson" : "{\n  \"model\" : \"granite4:3b\",\n  \"messages\" : [ {\n    \"role\" : \"user\",\n    \"content\" : \"Use your available tools to perform a long running operation for 2 seconds with 2 steps. DO NOT use any markdown formatting in the response.\"\n  }, {\n    \"role\" : \"assistant\",\n    \"tool_calls\" : [ {\n      \"function\" : {\n        \"name\" : \"longRunningOperation\",\n        \"arguments\" : {\n          \"duration\" : 2,\n          \"steps\" : 2\n        }\n      }\n    } ]\n  }, {\n    \"role\" : \"tool\",\n    \"content\" : \"Long running operation completed. Duration: 2 seconds, Steps: 2.\"\n  } ],\n  \"options\" : {\n    \"temperature\" : 0.3,\n    \"stop\" : [ ]\n  },\n  \"stream\" : false,\n  \"tools\" : [ {\n    \"type\" : \"function\",\n    \"function\" : {\n      \"name\" : \"longRunningOperation\",\n      \"description\" : \"Demonstrates a long running operation with progress updates\",\n      \"parameters\" : {\n        \"type\" : \"object\",\n        \"properties\" : {\n          \"duration\" : {\n            \"type\" : \"number\",\n            \"description\" : \"Duration of the operation in seconds\"\n          },\n          \"steps\" : {\n            \"type\" : \"number\",\n            \"description\" : \"Number of steps in the operation\"\n          }\n        },\n        \"required\" : [ ]\n      }\n    }\n  }, {\n    \"type\" : \"function\",\n    \"function\" : {\n      \"name\" : \"add\",\n      \"description\" : \"Adds two numbers\",\n      \"parameters\" : {\n        \"type\" : \"object\",\n        \"properties\" : {\n          \"a\" : {\n            \"type\" : \"number\",\n            \"description\" : \"First number\"\n          },\n          \"b\" : {\n            \"type\" : \"number\",\n            \"description\" : \"Second number\"\n          }\n        },\n        \"required\" : [ \"a\", \"b\" ]\n      }\n    }\n  }, {\n    \"type\" : \"function\",\n    \"function\" : {\n      \"name\" : \"echo\",\n      \"description\" : \"Echoes back the input\",\n      \"parameters\" : {\n        \"type\" : \"object\",\n        \"properties\" : {\n          \"message\" : {\n            \"type\" : \"string\",\n            \"description\" : \"Message to echo\"\n          }\n        },\n        \"required\" : [ \"message\" ]\n      }\n    }\n  } ]\n}",
+      "ignoreArrayOrder" : true,
+      "ignoreExtraElements" : false
+    } ]
+  },
+  "response" : {
+    "status" : 200,
+    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-03-03T15:26:02.290064872Z\",\"message\":{\"role\":\"assistant\",\"content\":\"The long running operation was executed successfully with a total duration of 2 seconds divided into 2 steps. The process completed as expected without any issues.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":2629078989,\"load_duration\":45655974,\"prompt_eval_count\":372,\"prompt_eval_duration\":868889130,\"eval_count\":31,\"eval_duration\":1689077612}",
+    "headers" : {
+      "Date" : "Tue, 03 Mar 2026 15:26:02 GMT",
+      "Content-Type" : "application/json; charset=utf-8"
+    }
+  },
+  "uuid" : "d366f73b-f02a-4f62-ad43-836f42226e05",
+  "persistent" : true,
+  "insertionIndex" : 12
+}
