+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-ea5aa343-879e-4dd0-94f5-c32886e33ec6.json
@@ -1,5 +1,5 @@
 {
-  "id" : "d9ade7be-8f94-4535-a552-9910f04cbbe0",
+  "id" : "ea5aa343-879e-4dd0-94f5-c32886e33ec6",
   "name" : "api_chat",
   "request" : {
     "url" : "/api/chat",
@@ -12,13 +12,13 @@
   },
   "response" : {
     "status" : 200,
-    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-01-21T08:01:45.122934296Z\",\"message\":{\"role\":\"assistant\",\"content\":\"That's a lovely choice of color! Blue is often associated with feelings of calmness and tranquility. If you need any assistance or have questions, feel free to ask.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":1661809217,\"load_duration\":41340745,\"prompt_eval_count\":69,\"prompt_eval_duration\":182241182,\"eval_count\":36,\"eval_duration\":1420810870}",
+    "body" : "{\"model\":\"granite4:3b\",\"created_at\":\"2026-03-03T15:26:24.28582676Z\",\"message\":{\"role\":\"assistant\",\"content\":\"That's a lovely choice of color! Blue is often associated with tranquility, stability, and depth. If you need any information or assistance related to the color blue or anything else, feel free to ask.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":2701949220,\"load_duration\":49357826,\"prompt_eval_count\":69,\"prompt_eval_duration\":378245954,\"eval_count\":43,\"eval_duration\":2245565434}",
     "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:01:45 GMT",
+      "Date" : "Tue, 03 Mar 2026 15:26:24 GMT",
       "Content-Type" : "application/json; charset=utf-8"
     }
   },
-  "uuid" : "d9ade7be-8f94-4535-a552-9910f04cbbe0",
+  "uuid" : "ea5aa343-879e-4dd0-94f5-c32886e33ec6",
   "persistent" : true,
   "insertionIndex" : 4
 }
