--- integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-f265676e-9355-40b6-9593-926440edd441.json
@@ -1,24 +1,0 @@
-{
-  "id" : "f265676e-9355-40b6-9593-926440edd441",
-  "name" : "api_chat",
-  "request" : {
-    "url" : "/api/chat",
-    "method" : "POST",
-    "bodyPatterns" : [ {
-      "equalToJson" : "{\n  \"model\" : \"orca-mini\",\n  \"messages\" : [ {\n    \"role\" : \"user\",\n    \"content\" : \"Describe the Miles of Camels Car Rental cancellations policy for cancelling 24 hours before pickup. What is the refund amount?\\n\\nAnswer using the following information:\\nCANCELLATION POLICY\\n- Cancellations made 24 hours before pickup: Full refund\\n- Cancellations made 12-24 hours before pickup: 50% refund\\n- Cancellations made less than 12 hours before pickup: No refund\\n\\nVEHICLE RETURN\\n- Vehicles must be returned with the same fuel level as at pickup.\\n- Late returns incur a fee of $25 per hour or fraction thereof.\\n\\nDAMAGE POLICY\\n- Minor damages under $200: Covered by insurance\\n- Major damages over $200: Customer responsibility\\n\\nMiles of Camels Car Rental - Company Information\\n\\nBUSINESS HOURS:\\nMonday-Friday: 8:00 AM - 6:00 PM\\nSaturday: 9:00 AM - 4:00 PM\\nSunday: Closed\\n\\nRENTAL AGREEMENT\\n- This agreement is between Miles of Camels Car Rental (\\\"Company\\\") and the customer (\\\"Renter\\\").\"\n  } ],\n  \"options\" : {\n    \"temperature\" : 0.3,\n    \"stop\" : [ ]\n  },\n  \"stream\" : false,\n  \"tools\" : [ ]\n}",
-      "ignoreArrayOrder" : true,
-      "ignoreExtraElements" : false
-    } ]
-  },
-  "response" : {
-    "status" : 200,
-    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-01-21T08:00:51.09034453Z\",\"message\":{\"role\":\"assistant\",\"content\":\" According to the Miles of Camels Car Rental cancellations policy, a full refund will be provided if the cancellation is made 24 hours or more before pickup. The refund amount will be the total cost of the rental including taxes and fees.\\n\\nAdditionally, according to the company's vehicle return policy, vehicles must be returned with the same fuel level as at pickup. Late returns incur a fee of $25 per hour or fraction thereof.\\n\\nIf a major damage occurs during the rental, the customer is responsible for the cost. If the damage is minor and covered by insurance, the cost will be reimbursed to the renter.\\n\\nIt's worth noting that Miles of Camels Car Rental business hours are Monday-Friday from 8:00 AM to 6:00 PM, Saturday from 9:00 AM to 4:00 PM, and Sunday is closed.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":13516197412,\"load_duration\":1071993252,\"prompt_eval_count\":314,\"prompt_eval_duration\":3516109695,\"eval_count\":191,\"eval_duration\":8893165055}",
-    "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:00:51 GMT",
-      "Content-Type" : "application/json; charset=utf-8"
-    }
-  },
-  "uuid" : "f265676e-9355-40b6-9593-926440edd441",
-  "persistent" : true,
-  "insertionIndex" : 21
-}
