+++ integration-tests/langchain4j-agent/src/test/resources/mappings/api_chat-f472d5b5-8789-4bd9-8e7d-8649badd493b.json
@@ -1,0 +1,27 @@
+{
+  "id" : "f472d5b5-8789-4bd9-8e7d-8649badd493b",
+  "name" : "api_chat",
+  "request" : {
+    "url" : "/api/chat",
+    "method" : "POST",
+    "bodyPatterns" : [ {
+      "equalToJson" : "{\n  \"model\" : \"orca-mini\",\n  \"messages\" : [ {\n    \"role\" : \"system\",\n    \"content\" : \"You are a whimsical storyteller. Your responses should be imaginative, descriptive, and always include a touch of magic. Start every story with 'Once upon a starlit night...'\"\n  }, {\n    \"role\" : \"user\",\n    \"content\" : \"Write a short story about a lost cat.\"\n  } ],\n  \"options\" : {\n    \"temperature\" : 0.3,\n    \"stop\" : [ ]\n  },\n  \"stream\" : false,\n  \"tools\" : [ ]\n}",
+      "ignoreArrayOrder" : true,
+      "ignoreExtraElements" : false
+    } ]
+  },
+  "response" : {
+    "status" : 200,
+    "body" : "{\"model\":\"orca-mini\",\"created_at\":\"2026-03-03T15:25:43.847297578Z\",\"message\":{\"role\":\"assistant\",\"content\":\" Once upon a starlit night, a curious little girl named Lily wandered through the forest, searching for her missing cat. She had been searching for hours and was about to give up when she heard a faint meowing sound. She followed the sound until she came across a small, shivering kitten.\\n\\nLily scooped up the tiny creature and took it back home with her. She fed it milk and gave it a warm bed to sleep in. Over time, the kitten grew stronger and more confident. It even started meowing again, this time with a new note of desperation.\\n\\nLily realized that the cat must have gone missing and was searching for its owner. She decided to use her magical powers to help find the cat's owner. She closed her eyes and made a wish, hoping it would lead her to the missing feline.\\n\\nWhen she opened her eyes, she saw a vision of a man walking his dog in the forest. The man was the cat's owner! Lily quickly called out to him and explained about the lost kitten. The man was relieved and grateful to have found his pet.\\n\\nLily used her magical powers once more to help the man locate the kitten's home. He searched high and low, but couldn't find it. Just as he was about to give up, he heard a faint meowing sound coming from inside the house. He rushed in and found the missing kitten, safe and sound.\\n\\nThe man was overjoyed and thanked Lily for her help. She replied with a smile, \\\"I did what any good fairy would do. I helped a lost little creature find its way home.\\\" And with that, Lily flew away, leaving the man and his dog to enjoy their reunited cat.\"},\"done\":true,\"done_reason\":\"stop\",\"total_duration\":22221023164,\"load_duration\":11261486,\"prompt_eval_count\":70,\"prompt_eval_duration\":1370467434,\"eval_count\":371,\"eval_duration\":20737438268}",
+    "headers" : {
+      "Date" : "Tue, 03 Mar 2026 15:25:43 GMT",
+      "Content-Type" : "application/json; charset=utf-8"
+    }
+  },
+  "uuid" : "f472d5b5-8789-4bd9-8e7d-8649badd493b",
+  "persistent" : true,
+  "scenarioName" : "scenario-1-api-chat",
+  "requiredScenarioState" : "Started",
+  "newScenarioState" : "scenario-1-api-chat-2",
+  "insertionIndex" : 16
+}
