--- integration-tests/langchain4j-agent/src/test/resources/mappings/api_embed-1e8084ee-91b3-4095-86b7-9845fe1480d6.json
@@ -1,24 +1,0 @@
-{
-  "id" : "1e8084ee-91b3-4095-86b7-9845fe1480d6",
-  "name" : "api_embed",
-  "request" : {
-    "url" : "/api/embed",
-    "method" : "POST",
-    "bodyPatterns" : [ {
-      "equalToJson" : "{\n  \"model\" : \"nomic-embed-text\",\n  \"input\" : [ \"Miles of Camels Car Rental - Company Information\\n\\nBUSINESS HOURS:\\nMonday-Friday: 8:00 AM - 6:00 PM\\nSaturday: 9:00 AM - 4:00 PM\\nSunday: Closed\\n\\nRENTAL AGREEMENT\\n- This agreement is between Miles of Camels Car Rental (\\\"Company\\\") and the customer (\\\"Renter\\\").\", \"RENTAL POLICIES:\\n- Minimum age: 21 years old\\n- Valid driver's license required\\n- Credit card required for security deposit\\n- Full tank of gas required at return\", \"VEHICLE FLEET:\\n- Economy cars: Starting at $29/day\\n- Mid-size cars: Starting at $39/day\\n- SUVs: Starting at $59/day\\n- Luxury vehicles: Starting at $89/day\", \"CANCELLATION POLICY\\n- Cancellations made 24 hours before pickup: Full refund\\n- Cancellations made 12-24 hours before pickup: 50% refund\\n- Cancellations made less than 12 hours before pickup: No refund\", \"VEHICLE RETURN\\n- Vehicles must be returned with the same fuel level as at pickup.\\n- Late returns incur a fee of $25 per hour or fraction thereof.\\n\\nDAMAGE POLICY\\n- Minor damages under $200: Covered by insurance\\n- Major damages over $200: Customer responsibility\", \"INSURANCE\\n- Basic insurance is included. Premium insurance available for $15/day.\\n\\nAGE REQUIREMENTS\\n- Minimum age: 21 years old\\n- Drivers under 25: Additional surcharge of $20/day\" ]\n}",
-      "ignoreArrayOrder" : true,
-      "ignoreExtraElements" : true
-    } ]
-  },
-  "response" : {
-    "status" : 200,
-    "bodyFileName" : "api_embed-1e8084ee-91b3-4095-86b7-9845fe1480d6.json",
-    "headers" : {
-      "Date" : "Wed, 21 Jan 2026 08:00:35 GMT",
-      "Content-Type" : "application/json; charset=utf-8"
-    }
-  },
-  "uuid" : "1e8084ee-91b3-4095-86b7-9845fe1480d6",
-  "persistent" : true,
-  "insertionIndex" : 23
-}
